Refactors the funometer routes to a separate admin_funometer_controller.
