+++ dashboard/app/views/admin_funometer/_funometer_levels.html.haml
@@ -1,18 +1,3 @@
-- content_for :head do
-  %script{src:'https://www.google.com/jsapi'}
-
-%h1
-  Fun-O-Meter Stats
-
-%h2
-  Stage: #{@stage_name}
-
-%h3
-  Overall Percentage:
-  = number_to_percentage(@overall_percentage, precision: 2)
-
-= render 'funometer_percentages_by_day'
-
 .by_level
   %h3
     Percentages by Level
