+++ dashboard/app/views/admin_funometer/_funometer_percentages_by_day.haml
