+++ dashboard/app/views/admin_funometer/funometer.html.haml
