+++ dashboard/app/views/admin_funometer/funometer_by_script.html.haml
@@ -32,25 +32,4 @@
         %td
           %span= stage.cnt
 
-.by_level
-  %h3
-    Percentages by Level
-  %table
-    %tr
-      - @ratings_by_level_headers.each do |header|
-        %th
-          %span= header
-    - @ratings_by_level.each do |level|
-      %tr
-        %td
-          %span
-            = link_to funometer_by_script_level_path(:script_id => @script_id, :level_id => level.level_id) do
-              = level.level_id
-        %td
-          %span
-            = link_to level_url(level.level_id) do
-              = level.name
-        %td
-          %span= number_to_percentage(level.percentage, precision: 2)
-        %td
-          %span= level.cnt
+= render 'funometer_levels'
