+++ dashboard/app/views/admin_funometer/funometer_by_script_level.html.haml
